feat: migrate components to app/lib/fps/_comps and update imports; add placeholders for server-side rendering
