+++ app/pages/fps.jsx
@@ -9,9 +9,9 @@ import PlayerLogic from './fps/_comps/PlayerLogic';
 import Floor from '../_components/objects/Floor';
 import Cube from '../_components/objects/Cube';
 import Crosshair from './fps/_comps/Crosshair';
-import Target from './fps/_comps/Target';
+import Target from '@/lib/fps/_comps/Target';
 import Bullet from './fps/_comps/Bullet';
-import Decal from './fps/_comps/Decal';
+import Decal from '@/lib/fps/_comps/Decal';
 import ShootingHandler from './fps/_comps/ShootingHandler';
 import PowerUp from './fps/_comps/PowerUp';
 import HillyFloor from '../_components/objects/HillyFloor';
